Added off arg for log level
