+++ dataspec/logging_handler.py
@@ -1,0 +1,30 @@
+"""
+Default module for configuring logging
+"""
+import logging
+import dataspec
+
+_MAPPING = {
+    "debug": logging.DEBUG,
+    "info": logging.INFO,
+    "warn": logging.WARN,
+    "error": logging.ERROR
+}
+
+
+@dataspec.registry.logging('default')
+def configure_logging(loglevel):
+    """
+    Default Handler for Logging Configuration
+    :param loglevel: loglevel as specified from the command line
+    :return: None
+    """
+    if loglevel.lower() in ['off', 'stop', 'disable']:
+        logging.disable(logging.CRITICAL)
+    else:
+        level = _MAPPING.get(loglevel, logging.INFO)
+        logging.basicConfig(
+            format='%(levelname)s [%(asctime)s] %(message)s',
+            level=level,
+            datefmt='%d-%b-%Y %I:%M:%S %p'
+        )
